// src/app/map/interfaces/stationInfo.interface.ts
export interface StationInfo {
  station_id_uuid: string;
  sensors:         Sensor[];
  generated_at:    number;
  station_id:      number;
}

export interface Sensor {
  lsid:                number;
  data:                Dato[];
  sensor_type:         number;
  data_structure_type: number;
}

export interface Dato {
  tz_offset:                          number;
  bar_trend_3_hr?:                    number;
  pressure_last?:                     number;
  ts:                                 number;
  bar?:                               number;
  bar_absolute?:                      number;
  bar_trend?:                         number;
  dew_point?:                         number;
  et_day?:                            number;
  forecast_rule?:                     number;
  forecast_desc?:                     string;
  heat_index?:                        number;
  hum_out?:                           number;
  rain_15_min_clicks?:                number;
  rain_15_min_in?:                    number;
  rain_15_min_mm?:                    number;
  rain_60_min_clicks?:                number;
  rain_60_min_in?:                    number;
  rain_60_min_mm?:                    number;
  rain_24_hr_clicks?:                 number;
  rain_24_hr_in?:                     number;
  rain_24_hr_mm?:                     number;
  rain_day_clicks?:                   number;
  rain_day_in?:                       number;
  rain_day_mm?:                       number;
  rain_rate_clicks?:                  number;
  rain_rate_in?:                      number;
  rain_rate_mm?:                      number;
  rain_storm_clicks?:                 number;
  rain_storm_in?:                     number;
  rain_storm_mm?:                     number;
  rain_storm_start_date?:             null;
  solar_rad?:                         number;
  temp_out?:                          number;
  thsw_index?:                        number;
  uv?:                                null;
  wind_chill?:                        number;
  wind_dir?:                          number;
  wind_dir_of_gust_10_min?:           number;
  wind_gust_10_min?:                  number;
  wind_speed?:                        number;
  wind_speed_2_min?:                  number;
  wind_speed_10_min?:                 number;
  wet_bulb?:                          number;
  iss_solar_panel_voltage?:           null;
  last_gps_reading_timestamp?:        number;
  resyncs?:                           null;
  transmitter_battery_state?:         number;
  crc_errors?:                        null;
  tiva_application_firmware_version?: number;
  lead_acid_battery_voltage?:         number;
  iss_transmitter_battery_voltage?:   null;
  beacon_interval?:                   number;
  davistalk_rssi?:                    null;
  solar_panel_voltage?:               number;
  rank?:                              number;
  false_wakeup_rssi?:                 number;
  cell_id?:                           number;
  longitude?:                         number;
  power_percentage_mcu?:              number;
  mcc_mnc?:                           number;
  iss_super_cap_voltage?:             null;
  false_wakeup_count?:                number;
  etx?:                               number;
  number_of_neighbors?:               number;
  last_parent_rtt_ping?:              number;
  bootloader_version?:                number;
  cme?:                               number;
  cc1310_firmware_version?:           null;
  power_percentage_rx?:               number;
  good_packet_streak?:                null;
  rpl_parent_node_id?:                null;
  afc_setting?:                       null;
  overall_access_technology?:         number;
  cell_channel?:                      number;
  noise_floor_rssi?:                  number;
  latitude?:                          number;
  cereg?:                             number;
  last_cme_error_timestamp?:          number;
  bluetooth_firmware_version?:        null;
  location_area_code?:                number;
  link_layer_packets_received?:       number;
  reception_percent?:                 null;
  rx_bytes?:                          number;
  link_uptime?:                       number;
  creg_cgreg?:                        number;
  health_version?:                    number;
  inside_box_temp?:                   number;
  tx_bytes?:                          number;
  elevation?:                         number;
  power_percentage_tx?:               number;
  rssi?:                              number;
  last_rx_rssi?:                      number;
  rpl_mode?:                          number;
  uptime?:                            number;
  platform_id?:                       number;
}